feat: deploy to sepolia and amoy
