+++ ignition/constructor-args/BurnMintTokenPool-4-9-2025_3-42-32_PM.js
@@ -1,0 +1,1 @@
+module.exports = ["0x8Ae0F967d59e2aEdF884248aF68bEc1932eb0435",18,[],"0xba3f6251de62dED61Ff98590cB2fDf6871FbB991","0x0BF3dE8c5D3e8A2B34D2BEeB17ABfCeBaf363A59"];
