+++ ignition/constructor-args/GovernanceSender-4-9-2025_3-43-44_PM.js
@@ -1,0 +1,1 @@
+module.exports = ["0x266f09eD17B7B7A8D1725Fd61B3935Ab02bA5f2b","0x0BF3dE8c5D3e8A2B34D2BEeB17ABfCeBaf363A59","0x779877A7B0D9E8603169DdbD7836e478b4624789"];
