+++ ignition/constructor-args/LeadrateSender-4-9-2025_3-43-44_PM.js
@@ -1,0 +1,1 @@
+module.exports = ["0xc6AEE9905648dB57ed700b1666650C5d8daE0d9F","0x0BF3dE8c5D3e8A2B34D2BEeB17ABfCeBaf363A59","0x779877A7B0D9E8603169DdbD7836e478b4624789"];
