+++ ignition/deployments/chain-11155111/deployed_addresses.json
@@ -1,0 +1,8 @@
+{
+  "CCIPPrepare#Frankencoin": "0x8Ae0F967d59e2aEdF884248aF68bEc1932eb0435",
+  "CCIPPrepare#BridgeAccounting": "0x3bF23a52590C2DF855bDbA748489C5fbBA54da12",
+  "CCIPPrepare#BurnMintTokenPool": "0xE9BD9035aAc87B84478C665f32FaAFA8858C626a",
+  "CCIPPrepare#CCIPAdmin": "0x39f6934D0C15Ef7bff883280aD97749DfBeaE8d5",
+  "CCIPPrepare#GovernanceSender": "0x08AdafD51205cEB550BB1b34769E89c8A609e6cA",
+  "CCIPPrepare#LeadrateSender": "0xB0351e929F6Bd42873Ec795919E0f03DbD797caa"
+}
