+++ scripts/deployment/deployments/amoy/.chainId
@@ -1,0 +1,1 @@
+80002
