+++ scripts/deployment/parameters/paramsCCIP.json
@@ -6,8 +6,8 @@
     "tokenAdminRegistry": "0x95F29FEE11c5C55d26cCcf1DB6772DE953B37B82",
     "rmnProxy": "0xba3f6251de62dED61Ff98590cB2fDf6871FbB991",
     "registryModuleOwner": "0x62e731218d0D47305aba2BE3751E7EE9E5520790",
-    "tokenPool": "0xd2bD88BBe0f63e4edFD1EcbD810446Ffd528c895",
-    "zchf": "0x93f3ACD0e1a63b87EEb8Ab9dA65178131fAd6b50",
+    "tokenPool": "0xE9BD9035aAc87B84478C665f32FaAFA8858C626a",
+    "zchf": "0x8Ae0F967d59e2aEdF884248aF68bEc1932eb0435",
     "chainSelector": "16015286601757825753"
   },
   {
@@ -17,13 +17,15 @@
     "tokenAdminRegistry": "0x1e73f6842d7afDD78957ac143d1f315404Dd9e5B",
     "chainSelector": "16281711391670634445",
     "mainnetChainSelector": "16015286601757825753",
-    "mainnetGovernanceSender": "0xa6B5a8dBeE81ADFC1144aD90332B36AC5Fcb8035",
+    "mainnetGovernanceSender": "0x08AdafD51205cEB550BB1b34769E89c8A609e6cA",
     "rmnProxy": "0x7c1e545A40750Ee8761282382D51E017BAC68CBB",
     "registryModuleOwner": "0x84ad5890A63957C960e0F19b0448A038a574936B",
-    "zchf": "0xE325c1dA55695c2a6347b6d4e8243eaBdBCddDD7",
-    "tokenPool": "0x5a803AeAAEa6304eAdae05Db19aeB18fbF15C5D5",
+    "zchf": "0x05C6eEb462A357Bf179e2a95e860eB000941EBaa",
+    "tokenPool": "0xcE2be1cC233802390ac77915529133758c0A8Ee3",
     "link": "0x0Fd9e8d3aF1aaee056EB9e802c3A762a667b1904",
-    "bridgeAccounting": "0x0000000000000000000000000000000000000000"
+    "bridgeAccounting": "0x3bF23a52590C2DF855bDbA748489C5fbBA54da12",
+    "mainnetLeadrateSender": "0xB0351e929F6Bd42873Ec795919E0f03DbD797caa",
+    "remoteChains": ["sepolia"]
   },
   {
     "chainName": "Local",
